adopt OOP for signature and http request
