+++ java/README.md
@@ -2,5 +2,5 @@
 Export the api key and secret key to your environment \
 `export BINANCE_API_KEY="<your api key>" && export BINANCE_API_SECRET="<your secret key>"` 
 
-`javac Spot.java`\
+`javac *.java`\
 `java Spot`
